added impl classes and constant class
