+++ src/main/java/edutech/backend/controller/CourseController.java
@@ -3,6 +3,7 @@ package edutech.backend.controller;
 import edutech.backend.entity.Course;
 import edutech.backend.exception.CourseNotFoundException;
 import edutech.backend.service.CourseService;
+import edutech.backend.util.MessageConstant;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,10 +27,10 @@ public class CourseController {
     @GetMapping("/{id}")
     public ResponseEntity<Course> getCourseById(@PathVariable Long id) {
         try {
-            Course course = courseService.findCourseById(id); // Calls the newly added method
+            Course course = courseService.findCourseById(id);
             return ResponseEntity.ok(course);
         } catch (CourseNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null); // Handle the exception
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
         }
     }
 
@@ -37,6 +38,32 @@ public class CourseController {
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Course> createCourse(@RequestBody Course course) {
         Course savedCourse = courseService.createCourse(course);
-        return ResponseEntity.status(HttpStatus.CREATED).body(savedCourse); // 201 Created
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedCourse);
+    }
+
+    @PutMapping("/{id}")
+    @PreAuthorize("hashRole('ADMIN')")
+    public ResponseEntity<String> updateCourse(@PathVariable Long id,@RequestBody Course updatedCourse){
+        try {
+            Course update = courseService.updateCourse(id,updatedCourse);
+            return ResponseEntity.status(HttpStatus.OK).body(MessageConstant.COURSE_WITH_ID+id+MessageConstant.UPDATED+MessageConstant.SUCCESSFULLY);
+        } catch (CourseNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.OK).body(e.getMessage());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(MessageConstant.AN_UNEXPECTED_ERROR_OCCURRED);
+        }
+    }
+
+    @DeleteMapping("/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<String>deleteCourse(@PathVariable Long id){
+        try {
+            courseService.deleteCourseById(id);
+            return ResponseEntity.status(HttpStatus.CREATED).body(MessageConstant.COURSE_WITH_ID +id+MessageConstant.DELETED+MessageConstant.SUCCESSFULLY);
+        }catch (CourseNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.OK).body(e.getMessage());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(MessageConstant.AN_UNEXPECTED_ERROR_OCCURRED);
+        }
     }
 }
