+++ src/main/java/edutech/backend/service/CustomUserDetailsService.java
@@ -1,34 +1,7 @@
 package edutech.backend.service;
 
-import edutech.backend.entity.User;
-import edutech.backend.repository.UserRepository;
-import edutech.backend.security.CustomUserDetails;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
 
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException {
-        // Check if the input is a username or email
-        User user;
-        if (usernameOrEmail.contains("@")) {
-            // If it contains '@', it's probably an email
-            user = userRepository.findByEmail(usernameOrEmail)
-                    .orElseThrow(() -> new UsernameNotFoundException("User not found with email: " + usernameOrEmail));
-        } else {
-            // Otherwise, it's treated as a username
-            user = userRepository.findByUsername(usernameOrEmail)
-                    .orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + usernameOrEmail));
-        }
-
-        return new CustomUserDetails(user);
-    }
+public interface CustomUserDetailsService {
+    UserDetails loadUserByUsername(String usernameOrEmail);
 }
