+++ src/main/java/edutech/backend/util/MessageConstant.java
@@ -1,0 +1,10 @@
+package edutech.backend.util;
+
+public class MessageConstant {
+    public static final String COURSE_WITH_ID = "Course with Id ";
+    public static final String DELETED = " Deleted";
+    public static final String SUCCESSFULLY = " Successfully";
+    public static final String AN_UNEXPECTED_ERROR_OCCURRED = "An unexpected error occurred.";
+    public static final String NOT_FOUND = " Not found.";
+    public static final String UPDATED = " Updated";
+}
